+++ .gitignore
@@ -333,3 +333,4 @@ ASALocalRun/
 # Local History for Visual Studio
 .localhistory/
 lib/
+Libraries/
