+++ BoostChat.sln
@@ -4,13 +4,8 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 VisualStudioVersion = 17.8.34525.116
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "Client", "Client\Client.vcxproj", "{9167BDEF-2746-4670-9BFD-B4E05A186F84}"
-	ProjectSection(ProjectDependencies) = postProject
-		{F885EA5A-73E5-44CF-AC22-D77CA53FBAB0} = {F885EA5A-73E5-44CF-AC22-D77CA53FBAB0}
-	EndProjectSection
 EndProject
 Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "Server", "Server\Server.vcxproj", "{041CEBB6-473D-481E-A584-A484A0939AEF}"
-EndProject
-Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "Common", "Common\Common.vcxproj", "{F885EA5A-73E5-44CF-AC22-D77CA53FBAB0}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -36,14 +31,6 @@ Global
 		{041CEBB6-473D-481E-A584-A484A0939AEF}.Release|x64.Build.0 = Release|x64
 		{041CEBB6-473D-481E-A584-A484A0939AEF}.Release|x86.ActiveCfg = Release|Win32
 		{041CEBB6-473D-481E-A584-A484A0939AEF}.Release|x86.Build.0 = Release|Win32
-		{F885EA5A-73E5-44CF-AC22-D77CA53FBAB0}.Debug|x64.ActiveCfg = Debug|x64
-		{F885EA5A-73E5-44CF-AC22-D77CA53FBAB0}.Debug|x64.Build.0 = Debug|x64
-		{F885EA5A-73E5-44CF-AC22-D77CA53FBAB0}.Debug|x86.ActiveCfg = Debug|Win32
-		{F885EA5A-73E5-44CF-AC22-D77CA53FBAB0}.Debug|x86.Build.0 = Debug|Win32
-		{F885EA5A-73E5-44CF-AC22-D77CA53FBAB0}.Release|x64.ActiveCfg = Release|x64
-		{F885EA5A-73E5-44CF-AC22-D77CA53FBAB0}.Release|x64.Build.0 = Release|x64
-		{F885EA5A-73E5-44CF-AC22-D77CA53FBAB0}.Release|x86.ActiveCfg = Release|Win32
-		{F885EA5A-73E5-44CF-AC22-D77CA53FBAB0}.Release|x86.Build.0 = Release|Win32
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
