+++ Server/App.cpp
@@ -4,7 +4,7 @@
 
 int main()
 {
-	boost::asio::io_context io_context;
+	std::shared_ptr<boost::asio::io_context> io_context = std::make_shared<boost::asio::io_context>();
 
 	TcpServer tcpServer(io_context, 4242);
 	tcpServer.Start();
