+++ Server/TcpSession.h
@@ -8,8 +8,8 @@ class TcpSession : public std::enable_shared_from_this<TcpSession>
 public:
 
 public:
-    TcpSession(boost::asio::io_context& io_context, ThreadSafeQueue<OwnedMessage>& qToServer)
-        : m_IoContext(io_context), m_Socket(io_context), m_QMessagesInServer(qToServer)
+    TcpSession(std::shared_ptr<boost::asio::io_context> io_context, ThreadSafeQueue<OwnedMessage>& qToServer)
+        : m_IoContext(io_context), m_Socket(*io_context), m_QMessagesInServer(qToServer)
     {
     }
     void Start();
@@ -27,11 +27,11 @@ private:
 
 private:
     boost::asio::ip::tcp::socket m_Socket;
-    boost::asio::io_context& m_IoContext;
+    std::shared_ptr<boost::asio::io_context> m_IoContext;
 
     Message m_TemporaryMessage;
 
-    ThreadSafeQueue<OwnedMessage>& m_QMessagesInServer;    // 서버로 보내는 메시지
-    ThreadSafeQueue<Message> m_QMessagesOutServer;    // 클라이언트로 보내는 메시지
+    ThreadSafeQueue<OwnedMessage>& m_QMessagesInServer;     // 서버로 보내는 메시지
+    ThreadSafeQueue<Message> m_QMessagesOutServer;          // 클라이언트로 보내는 메시지
 
 };
